Handle user not signed in like/unlike

+++ src/utils/blockstackStorage.js
@@ -3,32 +3,43 @@ import _ from "lodash";
 const appConfig = new AppConfig();
 const userSession = new UserSession({ appConfig });
 
+function userSignedIn() {
+  return userSession.isUserSignedIn();
+}
+
 export function putFile(path, data, options = {}, cb = () => {}) {
+  if (!userSignedIn()) return false;
   userSession.putFile(path, JSON.stringify(data), options).then(cb);
 }
 
 export async function appendToFile(path, data, options = {}, cb = () => {}) {
   let list = await readFile(path, options);
   if (!list) list = [];
-  userSession
-    .putFile(path, JSON.stringify(list.concat(data)), options)
-    .then(cb);
+  userSignedIn() &&
+    userSession
+      .putFile(path, JSON.stringify(list.concat(data)), options)
+      .then(cb);
 }
 
 export async function removeFromFile(path, value, options = {}, cb = () => {}) {
   let list = await readFile(path, options);
-  if (list) {
+  if (userSignedIn() && list) {
     userSession
-      .putFile(path, JSON.stringify(_.pull(list, value)), options)
+      .putFile(
+        path,
+        JSON.stringify(_.pullAllWith(list, [value], _.isEqual)),
+        options
+      )
       .then(cb);
   }
 }
 
 export async function readFile(path, options = {}) {
+  if (!userSignedIn()) return false;
   const file = await userSession.getFile(path, options);
   return JSON.parse(file);
 }
 
 export function deleteFile(path, options = {}, cb = () => {}) {
-  userSession.deleteFile(path).then(cb);
+  userSignedIn() && userSession.deleteFile(path).then(cb);
 }
